Major refactoring of LabelsInput component hierarchy
