+++ webpack.config.js
@@ -246,9 +246,13 @@ module.exports = (env, argv) => {
             mode,
             context,
             devtool,
+            resolve: {
+                extensions: ['.js', '.jsx']
+            },
             entry: glob
                 .sync('./widgets/common/src/props/*.js')
                 .concat(glob.sync('./widgets/common/src/hooks/*.js'))
+                .concat(glob.sync('./widgets/common/src/!(props|hooks)/*.js*'))
                 .concat(glob.sync('./widgets/common/src/*.js*')),
             output: {
                 path: path.join(outputPath, 'appData/widgets'),
